requirements.txt
torch>=1.9.0
torchaudio>=0.9.0
numpy>=1.19.2
pandas>=1.2.4
librosa>=0.8.1
pyyaml>=5.4.1
tqdm>=4.62.3
matplotlib>=3.4.3
epitran>=1.2
tensorboard>=2.7.0